(PC-28814) refactor(venueMap): export get geolocated venues in specific function
