+++ src/features/venueMap/hook/useGetVenuesInRegion.ts
@@ -1,7 +1,7 @@
 import { Venue } from 'features/venue/types'
 import { GeolocatedVenue } from 'features/venueMap/components/VenueMapView/types'
 import { calculateRoundRadiusInKilometers } from 'features/venueMap/helpers/calculateDistanceMap'
-import { isGeolocValid } from 'features/venueMap/helpers/isGeolocValid'
+import { getGeolocatedVenues } from 'features/venueMap/helpers/getGeolocatedVenues/getGeolocatedVenues'
 import { useGetAllVenues } from 'features/venueMap/useGetAllVenues'
 import { Region } from 'libs/maps/maps'
 
@@ -13,15 +13,5 @@ export const useGetVenuesInRegion = (
   const radius = calculateRoundRadiusInKilometers(region)
   const { venues = [] } = useGetAllVenues({ region, radius, initialVenues })
 
-  const geolocatedVenues = venues.filter(
-    (venue): venue is GeolocatedVenue => !!(venue.venueId && isGeolocValid(venue._geoloc))
-  )
-
-  const hasSelectionOutsideSearchArea =
-    selectedVenue && !geolocatedVenues.find((venue) => venue.venueId === selectedVenue.venueId)
-  if (hasSelectionOutsideSearchArea) {
-    geolocatedVenues.push(selectedVenue)
-  }
-
-  return geolocatedVenues
+  return getGeolocatedVenues(venues, selectedVenue)
 }
